requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
google-cloud-storage==2.18.2
google-cloud-documentai>=3.7.0,<4.0.0
google-cloud-discoveryengine==0.13.6
google-cloud-translate==3.15.3
google-auth==2.36.0
python-multipart==0.0.19
Pillow==10.4.0
opencv-python-headless==4.10.0.84
openai>=1.0.0,<2.0.0
python-dotenv
crawl4ai